requirements.txt
expects==0.9.0
flake8==3.8.4
flask==1.1.2
ipdb==0.13.4
sqlalchemy==1.3.20
elasticsearch==8.10.1